Функция по открытию приложения для авторизованного пользователя перенесена в отдельный класс TabBarViewControllerBuilder
