+++ Sprint2-Carkas/Sprint2-Carkas/MainViewController.swift
@@ -43,17 +43,7 @@ class MainViewController: UIViewController {
 //            navVCForReg.viewControllers = [regVC]
             present(navVCForReg, animated: true, completion: nil)
         } else {
-            let noteScreenVC = NoteScreenViewController()
-            let settingVC = SettingsViewController()
-            let tasksVC = TasksViewController()
-            let tabBar = UITabBarController()
-            let navVCForNotes = UINavigationController()
-            navVCForNotes.viewControllers = [noteScreenVC]
-            
-            
-            tabBar.viewControllers = [tasksVC, navVCForNotes, settingVC]
-            tabBar.modalPresentationStyle = .fullScreen
-            present(tabBar, animated: true, completion: nil)
+            present(TabBarViewControllerBuilder.makeAuthorizationAndOpenApp(), animated: true, completion: nil)
         }
         
     }
